feat(metrics): query toggle style
