+++ static/app/components/modals/metricWidgetViewerModal/queries.tsx
@@ -30,7 +30,7 @@ import {
   filterQueriesByDisplayType,
 } from 'sentry/views/dashboards/metrics/utils';
 import {DisplayType} from 'sentry/views/dashboards/types';
-import {EquationSymbol} from 'sentry/views/metrics/equationSymbol copy';
+import {EquationSymbol} from 'sentry/views/metrics/equationSymbol';
 import {FormulaInput} from 'sentry/views/metrics/formulaInput';
 import {getCreateAlert} from 'sentry/views/metrics/metricQueryContextMenu';
 import {QueryBuilder} from 'sentry/views/metrics/queryBuilder';
@@ -90,7 +90,6 @@ export function Queries({
               isHidden={query.isHidden}
               onChange={isHidden => onQueryChange({isHidden}, index)}
               disabled={!query.isHidden && visibleExpressions.length === 1}
-              isSelected={false}
               queryId={query.id}
               type={MetricExpressionType.QUERY}
             />
@@ -116,7 +115,6 @@ export function Queries({
               isHidden={equation.isHidden}
               onChange={isHidden => onEquationChange({isHidden}, index)}
               disabled={!equation.isHidden && visibleExpressions.length === 1}
-              isSelected={false}
               queryId={equation.id}
               type={MetricExpressionType.EQUATION}
             />
@@ -254,37 +252,35 @@ function EquationContextMenu({equationIndex, removeEquation}: EquationContextMen
 interface QueryToggleProps {
   disabled: boolean;
   isHidden: boolean;
-  isSelected: boolean;
   onChange: (isHidden: boolean) => void;
   queryId: number;
   type: MetricExpressionType;
 }
 
-function QueryToggle({
-  isHidden,
-  queryId,
-  disabled,
-  onChange,
-  isSelected,
-  type,
-}: QueryToggleProps) {
-  let tooltipTitle = isHidden ? t('Show query') : t('Hide query');
-  if (disabled) {
-    tooltipTitle = t('At least one query must be visible');
-  }
+function QueryToggle({isHidden, queryId, disabled, onChange, type}: QueryToggleProps) {
+  const tooltipTitle =
+    type === MetricExpressionType.QUERY
+      ? isHidden
+        ? t('Show query')
+        : t('Hide query')
+      : isHidden
+        ? t('Show equation')
+        : t('Hide equation');
 
   return (
-    <Tooltip title={tooltipTitle} delay={500}>
+    <Tooltip
+      title={!disabled ? tooltipTitle : t('At least one query must be visible')}
+      delay={500}
+    >
       {type === MetricExpressionType.QUERY ? (
         <StyledQuerySymbol
           isHidden={isHidden}
           queryId={queryId}
           isClickable={!disabled}
           aria-disabled={disabled}
-          isSelected={isSelected}
           onClick={disabled ? undefined : () => onChange(!isHidden)}
           role="button"
-          aria-label={isHidden ? t('Show query') : t('Hide query')}
+          aria-label={tooltipTitle}
         />
       ) : (
         <StyledEquationSymbol
@@ -292,10 +288,9 @@ function QueryToggle({
           equationId={queryId}
           isClickable={!disabled}
           aria-disabled={disabled}
-          isSelected={isSelected}
           onClick={disabled ? undefined : () => onChange(!isHidden)}
           role="button"
-          aria-label={isHidden ? t('Show query') : t('Hide query')}
+          aria-label={tooltipTitle}
         />
       )}
     </Tooltip>
@@ -320,11 +315,9 @@ const QueryWrapper = styled('div')<{hasQuerySymbol: boolean}>`
 `;
 
 const StyledQuerySymbol = styled(QuerySymbol)<{isClickable: boolean}>`
-  margin-top: 10px;
   ${p => p.isClickable && `cursor: pointer;`}
 `;
 const StyledEquationSymbol = styled(EquationSymbol)<{isClickable: boolean}>`
-  margin-top: 10px;
   ${p => p.isClickable && `cursor: pointer;`}
 `;
